fix: Attachments should not be deleted when their original document is deleted when referenced elsewhere

+++ server/api/utils.js
@@ -4,13 +4,14 @@ import debug from "debug";
 import Router from "koa-router";
 import { AuthenticationError } from "../errors";
 import { Document, Attachment } from "../models";
-import { Op } from "../sequelize";
+import { Op, sequelize } from "../sequelize";
+import parseAttachmentIds from "../utils/parseAttachmentIds";
 
 const router = new Router();
 const log = debug("utils");
 
 router.post("utils.gc", async (ctx) => {
-  const { token } = ctx.body;
+  const { token, limit = 500 } = ctx.body;
 
   if (process.env.UTILS_SECRET !== token) {
     throw new AuthenticationError("Invalid secret token");
@@ -25,23 +26,51 @@ router.post("utils.gc", async (ctx) => {
   };
 
   const documents = await Document.scope("withUnpublished").findAll({
-    attributes: ["id"],
+    attributes: ["id", "teamId", "text"],
     where,
+    paranoid: false,
+    limit,
   });
-  const documentIds = documents.map((d) => d.id);
 
-  await Attachment.destroy({
-    where: {
-      documentId: documentIds,
-    },
-  });
+  const query = `
+  SELECT COUNT(id)
+  FROM documents
+  WHERE "searchVector" @@ to_tsquery('english', :query) AND
+  "teamId" = :teamId AND
+  "id" != :documentId
+`;
+
+  for (const document of documents) {
+    const attachmentIds = parseAttachmentIds(document.text);
+
+    for (const attachmentId of attachmentIds) {
+      const [{ count }] = await sequelize.query(query, {
+        type: sequelize.QueryTypes.SELECT,
+        replacements: {
+          documentId: document.id,
+          teamId: document.teamId,
+          query: attachmentId,
+        },
+      });
+
+      if (parseInt(count) === 0) {
+        await Attachment.destroy({
+          where: {
+            id: attachmentId,
+          },
+        });
+
+        log(`Deleted attachment ${attachmentId}`);
+      }
+    }
+  }
 
   await Document.scope("withUnpublished").destroy({
     where,
     force: true,
   });
 
-  log(`Deleted ${documentIds.length} documents`);
+  log(`Deleted ${documents.length} documents`);
 
   ctx.body = {
     success: true,
